Add ability to equip/unequip weapon properly
Previously when we're double clicking the weapon item
from inventory, it only equips the weapon and no way
to unequip it from the back. Now we're adding an ability
to unequip weapons from the back simply by double clicking
on the weapon items in the inventory, or by using inventory
hotbar (1-5).

Signed-off-by: MOVZX <movzx@yahoo.com>